# docs/05-api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Opportunity Intelligence Platform — Public API
  version: 1.0.0
  description: >
    Read-first public API over the platform's shared intelligence (trends, scores, entities,
    search) plus tenant resources (watchlists, alerts, reports). Auth via scoped, rate-limited,
    metered API keys. Errors follow RFC 9457 (application/problem+json). This is a baseline spec;
    the source of truth is generated from the Fastify route schemas in services/api.
servers:
  - url: https://api.aioi.example/api/v1
security:
  - ApiKeyAuth: []
tags:
  - name: Trends
  - name: Opportunities
  - name: Entities
  - name: Search
  - name: Watchlists
  - name: Alerts
paths:
  /trends:
    get:
      tags: [Trends]
      summary: List trends
      parameters:
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, maximum: 100, default: 25 } }
        - {
            name: status,
            in: query,
            schema: { type: string, enum: [EARLY, ACTIVE, FADING, ARCHIVED] },
          }
        - { name: topic, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string, example: "-lastSignalAt" } }
      responses:
        "200":
          description: Paged list of trends
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/TrendSummary" } }
                  nextCursor: { type: string, nullable: true }
        "429": { $ref: "#/components/responses/RateLimited" }
  /trends/{id}:
    get:
      tags: [Trends]
      summary: Get trend detail (scores + action plan + evidence)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Trend detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrendDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
  /search:
    get:
      tags: [Search]
      summary: Keyword or semantic search
      parameters:
        - { name: q, in: query, required: true, schema: { type: string } }
        - {
            name: mode,
            in: query,
            schema: { type: string, enum: [keyword, semantic], default: keyword },
          }
      responses:
        "200":
          description: Ranked results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/TrendSummary" } }
  /watchlists:
    get:
      tags: [Watchlists]
      summary: List watchlists (tenant-scoped)
      responses:
        "200": { description: OK }
    post:
      tags: [Watchlists]
      summary: Create a watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, workspaceId]
              properties:
                name: { type: string }
                workspaceId: { type: string, format: uuid }
      responses:
        "201": { description: Created }
        "422": { $ref: "#/components/responses/ValidationError" }
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
  responses:
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ValidationError:
      description: Input validation failed
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
    Score:
      type: object
      properties:
        dimension:
          type: string
          enum:
            [
              opportunity,
              business,
              developer,
              creator,
              seo,
              competition,
              monetization,
              risk,
              difficulty,
              predictedLifetime,
            ]
        value: { type: integer, minimum: 0, maximum: 100 }
        band: { type: string, enum: [low, medium, high] }
        confidence: { type: number, minimum: 0, maximum: 1 }
        rationale: { type: string }
        evidence: { type: array, items: { type: string } }
        rubricVersion: { type: string }
    TrendSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        title: { type: string }
        status: { type: string }
        opportunityScore: { type: integer }
    TrendDetail:
      allOf:
        - $ref: "#/components/schemas/TrendSummary"
        - type: object
          properties:
            summary: { type: string }
            scores: { type: array, items: { $ref: "#/components/schemas/Score" } }
            actionPlan: { type: object, additionalProperties: true }
            entities: { type: array, items: { type: object } }
            signals: { type: array, items: { type: object } }